test: fixes
