+++ tests/unit/schemas/custom_id_field_name.py
@@ -2,8 +2,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from strawchemy.config import StrawchemyConfig
-from strawchemy.mapper import Strawchemy
+from strawchemy import Strawchemy, StrawchemyConfig
 
 import strawberry
 from strawberry import Info, auto
